requirements.txt
asgiref==3.6.0
contextlib2==21.6.0
Django==4.2
django-ninja==0.21.0
django-ninja-extra==0.18.7
injector==0.20.1
pydantic==1.10.7
sqlparse==0.4.4
typing_extensions==4.5.0
